requirements.txt
python-dotenv
langchain
langchain-groq
langchain-community
fastapi
uvicorn
qdrant-client